Rename to superstack
Make it work with the new openstack client tool set

This is a bit of a departure from the supernova basis from which this is
forked. It does not look to use any existing supernova setup, and
instead will require fresh configuration, which may be an opportunity to
clean up cruft. A dependency on python-openstackclient has been added to
help get up and going.

http://docs.openstack.org/developer/python-openstackclient